feat(badge): change default size to lg
The default size for the Badge component has been updated from `md` to `lg` to provide a better default visual appearance.

The documentation for the Badge and Tooltip components has been updated to reflect this new default.

BREAKING CHANGE: The default size of the Badge component is now 'lg' instead of 'md'.